Filter dropdown cleanup
